Make kafka integration more testable
