+++ core/src/commands/impl/play.spec.ts
@@ -11,6 +11,6 @@ test('should not create player if previously created', () => {
   const play = createTestEvent({ content: '/play test' })
   const playAgain = createTestEvent({ content: '/play test2' })
 
-  const messages = handle(hot('ab|', { a: play, b: playAgain }))
+  const messages = handle(hot('ab', { a: play, b: playAgain }))
   expect(messages).toMatchSnapshot()
 })
