+++ core/src/jevent/Factory.ts
@@ -1,6 +1,6 @@
 import * as Player from '../player/Player'
 import { delayFactoryImpl } from './impl/delay'
-import { kafkaFactoryImpl } from './impl/kafka'
+import { kafkaProducerFactoryImpl } from './impl/kafka'
 import { timeoutFactoryImpl } from './impl/timeout'
 import { UUIDFactoryImpl } from './impl/uuid'
 import { ytSearchFactoryImpl } from './impl/ytSearch'
@@ -12,7 +12,7 @@ export interface Factory {
     readonly timeout: TimeoutFactory
     readonly ytSearch: YtSearchFactory
     readonly uuid: UUIDFactory
-    readonly kafka: KafkaFactory
+    readonly kafkaProducer: KafkaProducerFactory
   }
 }
 
@@ -24,4 +24,4 @@ export type YtSearchFactory = typeof ytSearchFactoryImpl
 
 export type UUIDFactory = typeof UUIDFactoryImpl
 
-export type KafkaFactory = typeof kafkaFactoryImpl
+export type KafkaProducerFactory = typeof kafkaProducerFactoryImpl
