+++ core/src/main.ts
@@ -1,8 +1,8 @@
 import { Client, Intents, Message } from 'discord.js'
 import { fromEvent, Observable } from 'rxjs'
 import apolloInit from './apollo/apollo'
+import { handleDisconnects } from './commands/impl/Disconnect'
 import config from './config'
-import { handleDisconnects } from './disconnect'
 import * as Event from './jevent/JEvent'
 import logger from './logger'
 import { handleMessageEvents, sendMessageEvent } from './messageHandler'
