+++ core/src/playlist/playlistTracker.spec.ts
@@ -1,4 +1,5 @@
-import { map, Observable } from 'rxjs'
+import { map, merge, Observable } from 'rxjs'
+import { disconnectFromChannel } from '../commands/impl/Disconnect'
 import { handleMessage } from '../messageHandler'
 import { createTestEvent, hot, sandbox } from '../test/util'
 import { TRACKING_EVENT } from './PlaylistTracker'
@@ -6,7 +7,8 @@ import { TRACKING_EVENT } from './PlaylistTracker'
 test('should create tracker for playlist', () => {
   const play = createTestEvent({ content: '/play test' })
 
-  const messages = handleTrackingMessages(hot('a', { a: play }))
+  const messages = sandbox.getMessages(handleTrackingMessages(hot('a', { a: play })))
+
   expect(messages).toMatchSnapshot()
 })
 
@@ -14,7 +16,8 @@ test('should track multiple tracks in the same playlist', () => {
   const play = createTestEvent({ content: '/play test' })
   const play2 = createTestEvent({ content: '/play test2' })
 
-  const messages = handleTrackingMessages(hot('ab', { a: play, b: play2 }))
+  const messages = sandbox.getMessages(handleTrackingMessages(hot('ab', { a: play, b: play2 })))
+
   expect(messages).toMatchSnapshot()
 })
 
@@ -22,11 +25,15 @@ test('should differentiate tracking between playlists', () => {
   const play = createTestEvent({ content: '/play test' })
   const disconnect = createTestEvent({ content: '/disconnect' })
 
-  const messages = handleTrackingMessages(hot('aba', { a: play, b: disconnect }))
-  expect(messages).toMatchSnapshot()
+  const messages = handleTrackingMessages(hot('a-a', { a: play }))
+  const messages2 = disconnectFromChannel(hot('-b-', { b: disconnect })).pipe(
+    map(msg => msg.result),
+  )
+
+  expect(sandbox.getMessages(merge(messages, messages2))).toMatchSnapshot()
 })
 
-const handleTrackingMessages = (source: Observable<any>) => sandbox.getMessages(handleMessage(source).pipe(
+const handleTrackingMessages = (source: Observable<any>) => handleMessage(source).pipe(
   map(msg => {
     // include the actual item when matching tracking events
     if (msg.result === TRACKING_EVENT) {
@@ -35,4 +42,4 @@ const handleTrackingMessages = (source: Observable<any>) => sandbox.getMessages(
       return msg.result
     }
   }),
-))
+)
